# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 명지피앤피 인사·총무 DX 시스템 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalEmployees: number;
  foreignEmployees: number;
  visaExpiringSoon: number;
  educationUpcoming: number;
  insuranceRenewalSoon: number;
  isoScheduleCount: number;
}

export interface Employee {
  id: number;
  employeeNumber: string;
  name: string;
  department: string;
  position: string;
  hireDate: string;
  /** @nullable */
  phone?: string | null;
  isForeigner: boolean;
  /** @nullable */
  visaType?: string | null;
  /** @nullable */
  visaExpiryDate?: string | null;
  createdAt: string;
}

export interface EmployeeDetail {
  id: number;
  employeeNumber: string;
  name: string;
  department: string;
  position: string;
  hireDate: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  birthDate?: string | null;
  isForeigner: boolean;
  /** @nullable */
  nationality?: string | null;
  /** @nullable */
  visaType?: string | null;
  /** @nullable */
  visaExpiryDate?: string | null;
  /** @nullable */
  passportExpiryDate?: string | null;
  /** @nullable */
  alienRegistrationExpiryDate?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface EmployeeInput {
  employeeNumber: string;
  name: string;
  department: string;
  position: string;
  hireDate: string;
  phone?: string;
  email?: string;
  address?: string;
  birthDate?: string;
  isForeigner?: boolean;
  nationality?: string;
  visaType?: string;
  visaExpiryDate?: string;
  passportExpiryDate?: string;
  alienRegistrationExpiryDate?: string;
  notes?: string;
}

export interface PersonnelHistory {
  id: number;
  employeeId: number;
  /** 입사|승진|부서이동|직책변경 */
  type: string;
  date: string;
  description: string;
  /** @nullable */
  previousDepartment?: string | null;
  /** @nullable */
  newDepartment?: string | null;
  /** @nullable */
  previousPosition?: string | null;
  /** @nullable */
  newPosition?: string | null;
  createdAt: string;
}

export interface PersonnelHistoryInput {
  type: string;
  date: string;
  description: string;
  previousDepartment?: string;
  newDepartment?: string;
  previousPosition?: string;
  newPosition?: string;
}

export interface Education {
  id: number;
  employeeId: number;
  name: string;
  date: string;
  completed: boolean;
  /** @nullable */
  certificateFile?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface EducationInput {
  name: string;
  date: string;
  completed: boolean;
  certificateFile?: string;
  notes?: string;
}

export interface Reward {
  id: number;
  employeeId: number;
  /** 포상|징계 */
  type: string;
  date: string;
  content: string;
  createdAt: string;
}

export interface RewardInput {
  type: string;
  date: string;
  content: string;
}

export interface Disciplinary {
  id: number;
  employeeId: number;
  disciplinaryType: string;
  date: string;
  content: string;
  createdAt: string;
}

export interface DisciplinaryInput {
  disciplinaryType: string;
  date: string;
  content: string;
}

export interface Interview {
  id: number;
  employeeId: number;
  date: string;
  content: string;
  /** @nullable */
  interviewer?: string | null;
  createdAt: string;
}

export interface InterviewInput {
  date: string;
  content: string;
  interviewer?: string;
}

export interface ForeignerInfo {
  id: number;
  employeeId: number;
  visaType: string;
  visaExpiryDate: string;
  passportExpiryDate: string;
  alienRegistrationExpiryDate: string;
  /** @nullable */
  notes?: string | null;
  updatedAt: string;
}

export interface ForeignerInfoInput {
  visaType: string;
  visaExpiryDate: string;
  passportExpiryDate: string;
  alienRegistrationExpiryDate: string;
  notes?: string;
}

export interface Attachment {
  id: number;
  employeeId: number;
  /** 근로계약서|자격증|교육수료증|징계문서|기타 */
  category: string;
  fileName: string;
  fileUrl: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
}

export interface AttachmentInput {
  category: string;
  fileName: string;
  fileUrl: string;
  fileSize?: number;
}

export interface TimelineEvent {
  date: string;
  type: string;
  description: string;
  category: string;
}

export interface VisaSchedule {
  employeeId: number;
  employeeName: string;
  department: string;
  visaType: string;
  expiryDate: string;
  daysUntilExpiry: number;
  /** danger|warning|info */
  urgency: string;
}

export interface EducationSchedule {
  id: number;
  employeeId: number;
  employeeName: string;
  department: string;
  educationName: string;
  scheduledDate: string;
  daysUntil: number;
  urgency: string;
}

export interface IsoSchedule {
  id: number;
  title: string;
  scheduledDate: string;
  type: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface IsoScheduleInput {
  title: string;
  scheduledDate: string;
  type: string;
  description?: string;
}

export interface InsuranceSchedule {
  id: number;
  insuranceName: string;
  renewalDate: string;
  /** @nullable */
  insurer?: string | null;
  /** @nullable */
  amount?: number | null;
  daysUntilRenewal: number;
  urgency: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface InsuranceScheduleInput {
  insuranceName: string;
  renewalDate: string;
  insurer?: string;
  amount?: number;
  notes?: string;
}

export interface Document {
  id: number;
  title: string;
  /** 근로계약서|자격증|교육자료|ISO|ESG|징계문서|기타 */
  category: string;
  /** @nullable */
  employeeId?: number | null;
  /** @nullable */
  employeeName?: string | null;
  fileName: string;
  fileUrl: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
}

export interface DocumentInput {
  title: string;
  category: string;
  employeeId?: number;
  fileName: string;
  fileUrl: string;
  fileSize?: number;
}

export type ListEmployeesParams = {
name?: string;
department?: string;
position?: string;
};

export type ListDocumentsParams = {
category?: string;
employeeName?: string;
};