# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 명지피앤피 인사·총무 DX 시스템 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: 문서 목록 조회
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: employeeName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: 문서 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: 문서 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
          description: "근로계약서|자격증|교육자료|ISO|ESG|징계문서|기타"
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: ["integer", "null"]
        uploadedAt:
          type: string
      required:
        - id
        - title
        - category
        - fileName
        - fileUrl
        - uploadedAt
    DocumentInput:
      type: object
      properties:
        title:
          type: string
        category:
          type: string
        employeeId:
          type: integer
        fileName:
          type: string
        fileUrl:
          type: string
        fileSize:
          type: integer
      required:
        - title
        - category
        - fileName
        - fileUrl