# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * 명지피앤피 인사·총무 DX 시스템 API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  title: string;
  /** 근로계약서|자격증|교육자료|ISO|ESG|징계문서|기타 */
  category: string;
  /** @nullable */
  employeeId?: number | null;
  /** @nullable */
  employeeName?: string | null;
  fileName: string;
  fileUrl: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
}